requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.2.1
fastapi==0.115.12
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.5
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
httpx>=0.25.0
Pillow>=9.0.0
python-multipart>=0.0.5
aiofiles>=22.0.0
jinja2>=3.0.0
python-jose>=3.3.0
passlib>=1.7.0
uvicorn>=0.18.0
aiosqlite>=0.17.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
aiosmtplib>=1.1.0
